Did some madddd changes to loading & routing
Prevent children from rendering before jwt token is validated

JWT Token validation is asynchronous and is a non blocking function so the main thread of execution will carry on while async completes and resolves its promise, allowing the result to be fed back to the event loop later on. This allows for concurrent execution, and prevents laggy UI, as one query which takes long to execute may not allow the remaining code to execute.

This effectively causes the validation to be pushed out of main line of execution, and the code will render the children first. In this case the sign in page first. Why would we want sign in page to be there when the jwt token already exists, and could be valid, in which case we should be directing user to /questions route immediately without them noticing the rendering of the sign in page?

So by having the loadingInitial, we will force the children to not be rendered before the token has been validated, and this logic and implementation can be found under AuthContext.js :+1: